scripts/fontlist/requirements.txt
# requirements.txt
fonttools
pytest  # Include pytest for testing
python-dotenv
